Add delimLength field for proper calculation of multiple of 3 rule.
The rule requiring that, if either open or close delimiter run
can both open and close, the sum of the lengths of these runs
not be a multiple of 3, refers to the full length of the
delimiter runs.  We were only paying attention to the current
length (with "used" delims subtracted).

Caught by a regression test.