Add nonbacktracking state to inline parser.
The point of this is to store information about when we've
scanned to the end of the input looking for an "ender"
(e.g. for CDATA or <?) to prevent duplicate parsing and
quadratic performance in pathological cases.

This comes with a ~5% performance hit, probably because of
the overhead from the monad transformer.

Closes #7.