Add advice to haddocks for pipeTableSpec.
Partially addresses #52.

The problem is that if a line could be a candidate pipe
table heading, but the following line of separators is
not encountered, the line is treated as a paragraph, even
if it has indications of other block-level formatting.

Putting the pipeTableSpec AFTER parsers for lists, headings,
etc. causes the latter to take priority.