bracketedSuffix now takes a [Tok] rather than Text parameter.
[API change].  This parameter is used for the reference key
in shortcut and compact reference links. For that purpose a
Text was sufficient, but for other parsers involving matched
brackets, it may be useful to recover the original tokens
inside the brackets (e.g. for reparsing).

Note that this is a breaking change to the API and could
potentially break some third-party extensions.