Changed type of bracketedSuffix.
Instead of a `[Tok]`, we now pass it a `[([Tok], il)]`, with chunks
of tokens associated with parsed inlines.

This gives the parser flexibility to break up this content
and analyze it.  The raw tokens can still be extracted with
`concatMap fst`.

Add `chunksToPairs`, define `unChunks` in terms of it.