Change mkFormattingSpecMap so it integrates different...
FormattingSpecs that use the same character.  Otherwise
we have problems if you have one formatting spec that
reacts to single delimiters and another that reacts to
pairs; if the first fails to match, the fallback behavior
is produced and the second never matches.  Closes #87.

I think this points the way to a better overall architecture,
but it's worth locking in this change now.