Inlines: Add bracketedSplitOn and change type of...
bracketedSuffix.  We go back to having a Text as the "raw key"
parameter.  But now the constructor operates on a list of
il, to be split up on the (unescaped) character (if any) specified in
bracketedSplitOn.

Note that emphasis is resolved within the split-up chunks,
and won't cross the split character.