commonmark: re-export Text.Parsec.Pos accessors
To deal robustly with source positions of document elements (e.g.
for deletion or alteration) we need to be able to drill into the
SourcePos values in the SourceRange.  It is a bit of a nuisance to
require users to have a direct dependency on *parsec* just to get at
`sourceLine` and `sourceColumn`.  Re-export them from
`Commonmark.Types`.

Fixes: https://github.com/jgm/commonmark-hs/issues/106