Fix pipe table parser interfering with other block structures
This parser is structured as a system that parses the *second* line first,
then parses the first line. That is, if it detects a delimiter row as the
second line of a paragraph, it converts the paragraph into a table. This seems
counterintuitive, but it works better than trying to convert a table into
a paragraph, since it might need to be something else.

As a comparison with GitHub:

  * GitHub allows tables to interrupt paragraphs, but this parser doesn't.
    The old parser didn't used to support this, either, so it's not
    a regression.
  * The odd behavior where tables don't get lazily continued is actually
    the same thing GitHub and pulldown-cmark both do.