Make pipe_tables extension treat backslash escapes like GH does (#112)
This change essentially changes the way the text `\\|` gets parsed inside a table.
In the old version, the first backslash escapes the second backslash, and then
the pipe is treated as a cell separator. In the new version, the pipe is still preceded
by a backslash, so it is still literal text.

This change seems reasonable, because if you want the pipe to act as a separator,
you can always separate it from the cell contents with a space in the new system,
but if you want to write the literal text `\|` inside a cell in the old system, you can't.

The escaping rule is documented in detail in the spec for this extension.

This change also aligns our escaping of pipes with GitHub's.