Avoid adding trailing newline to list block if it's already there
This can happen because of the two-line limit on blank list items,
as seen in this markdown sample, marked up with line numbers

```text
 1 | * x
 2 | * -
   |     ^ first blank line of ListItem
 3 |
   |     ^ second, final blank line of ListItem
 4 |
   |   ^ blank line is part of List, not ListItem
 5 | Test
```

The list item has `listItemBlanksAtEnd = True`, because of line 3,
but line 4 gets added to the List's blockBlanks. This means
blockFinalize should avoid adding it, otherwise you get
`blockBlanks' = [4, 4]` and that causes it to spurriously be
parsed as loose.