Improve autolinks extension.
The autolinks extension was interacting badly with explicit links,
as shown by #147:

    [link](https://baidu.com)aaa<span></span>bbb

To fix this we had to make autolink parsing a bit stricter than
the GFM spec does.  They allow unbalanced `)` except at the end
of a URL (which is defined as: followed by optional final punctuation
then whitespace or eof).  With this change, we don't allow unbalanced
`)` at all in raw URLs. This should not be a big problem in practice.

Closes #147.